move MessageDialog to atoms
