+++ src/components/organisms/FactionListActionToolbar/FactionListActionToolbar.tsx
@@ -5,8 +5,8 @@ import { GameContext } from "../../../contexts/GameContext";
 import { useSelection } from "../../../hooks/useSelection";
 import AddFactionDialog from "../../molecules/AddFactionDialog";
 import ListActionToolbar from "../../molecules/ListActionToolbar";
-import MessageDialog from "../../molecules/MessageDialog";
-import { DialogActionHandler } from "../../molecules/MessageDialog/MessageDialog";
+import MessageDialog from "../../atoms/MessageDialog";
+import { DialogActionHandler } from "../../atoms/MessageDialog/MessageDialog";
 
 export default function FactionListActionToolbar() {
   const { controller } = useContext(GameContext);
