+++ src/components/organisms/LocationsActionToolbar/LocationsActionToolbar.tsx
@@ -6,8 +6,8 @@ import { useSelection } from "../../../hooks/useSelection";
 import LocationInfo from "../../../types/LocationInfo";
 import AddLocationDialog from "../../molecules/AddLocationDialog";
 import ListActionToolbar from "../../molecules/ListActionToolbar";
-import MessageDialog from "../../molecules/MessageDialog";
-import { DialogActionHandler } from "../../molecules/MessageDialog/MessageDialog";
+import MessageDialog from "../../atoms/MessageDialog";
+import { DialogActionHandler } from "../../atoms/MessageDialog/MessageDialog";
 
 export default function LocationsActionToolbar() {
   const { controller } = useContext(GameContext);
